fix: resolve PR review comments
- Move RelatedProposalsQuery and RelatedProposalsResponse structs to models directory
- Refactor Exa service to use project error types instead of anyhow
- Simplify extract_domain function using unwrap_or pattern
- Simplify ok_or_else to ok_or in handlers.rs
- Fix all clippy warnings and ensure code quality
